修复之前版本存在的大量BUG
1. （暂时）无效化运算后按运算符号键
2. 运算后直接按数字键会自动触发清屏
3. 结果支持浮点数
4. 运算中按运算符号键会直接修改算式中的符号
5. 直接在算式中按运算符号键会直接将第一个数设置为0
6. 按完运算符号键后直接按运算键会直接将第二个数设置为0
7. 增加对0键在值为0时的大量特判
8. 历史窗口及导入、导出历史窗口按钮按下后会使按钮临时失效，直至打开的窗口关闭

+++ src/main/java/com/wyd/julyed/tool/ImportDocument.java
@@ -26,7 +26,7 @@ public class ImportDocument {
         calcuation.setParameter1(Integer.valueOf(item.getAttributeValue(Constant.STRING_PARAMETER1)));
         calcuation.setOperator(Operator.getOperator(item.getAttributeValue(Constant.STRING_OPERATOR)));
         calcuation.setParameter2(Integer.valueOf(item.getAttributeValue(Constant.STRING_PARAMETER2)));
-        calcuation.setResult(Integer.valueOf(item.getAttributeValue(Constant.STRING_RESULT)));
+        calcuation.setResult(Double.valueOf(item.getAttributeValue(Constant.STRING_RESULT)));
         GlobalManager.getList().add(calcuation);
     }
 }
